backend/pyproject.toml
[project]
name = "charon-backend"
version = "0.0.8"
description = "3D dependency visualizer for Python projects - Backend API"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Charon Contributors" }
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "networkx>=3.2.1",
    "numpy>=1.24.0",
    "aiohttp>=3.9.1",
    "httpx-oauth>=0.15.0",
    "python-multipart>=0.0.6",
    "toml>=0.10.2",
    "sse-starlette>=1.8.2",
    "radon>=6.0.1",
    "weasyprint>=63.0",
    "pyyaml>=6.0.1",
    "tree-sitter-language-pack>=0.13.0",
    "cachetools>=6.2.2",
    "jinja2>=3.1.6",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
